Add files via upload
This notebook adress the problem of predicting if the passangers of the Titanic would survive or not, using some simple information about demography, tickets and location. In this case, the whole process to address a Data Science process is done. Starting by the cleaning of the data, the exploratory data análisis, feature engineering and implementation and tuning of models.